requirements.txt
astroid==2.4.2
cheroot==8.4.5
CherryPy==18.6.0
importlib-resources==3.0.0
isort==5.6.4
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
numpy==1.15.4
PageCalc==0.2.3
pandas==0.23.4
portend==2.6
pylint==2.6.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
secure==0.1.4
six==1.15.0
SQLAlchemy==1.3.20
tempora==4.0.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zc.lockfile==2.0
zipp==3.3.0
